+++ neighbor_board.h
@@ -5,7 +5,7 @@
 #include <stdint.h>
 
 void neighbor_init(void);
-uint8_t neighbor_at(int x, int y);
-void neighbor_set(int x, int y, uint8_t value);
+int neighbor_at(int x, int y);
+void neighbor_set(int x, int y, int value);
 
 #endif
